pyproject.toml
[project]
name = "cinema"
version = "0.1.0"
description = ""
authors = [
    {name = "amitavaghosh2",email = "amitava@rendernet.ai"}
]
readme = "README.md"
requires-python = "<3.13,>=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.9.0",
    "crewai[tools] (>=1.3.0,<2.0.0)",
    "google-genai>=1.48.0",
    "imagehash>=4.3.2",
    "isort>=7.0.0",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "networkx[default]>=3.5",
    "pyyaml>=6.0.3",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
